Move converters mapping table into Configuration namespace.
