+++ nanoFramework.Json.Test/Configuration/ConvertersMappingTests.cs
@@ -1,9 +1,10 @@
-﻿using nanoFramework.Json.Converters;
+﻿using nanoFramework.Json.Configuration;
+using nanoFramework.Json.Converters;
 using nanoFramework.TestFramework;
 using System;
 using System.Text;
 
-namespace nanoFramework.Json.Test.Converters
+namespace nanoFramework.Json.Test.Configuration
 {
     [TestClass]
     public class ConvertersMappingTests
