+++ nanoFramework.Json.Test/nanoFramework.Json.Test.nfproj
@@ -32,7 +32,7 @@
   </PropertyGroup>
   <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.props" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.props')" />
   <ItemGroup>
-    <Compile Include="Converters\ConvertersMappingTests.cs" />
+    <Compile Include="Configuration\ConvertersMappingTests.cs" />
     <Compile Include="Converters\SByteConverterTests.cs" />
     <Compile Include="Converters\ByteConverterTests.cs" />
     <Compile Include="Converters\BoolConverterTests.cs" />
@@ -46,7 +46,7 @@
     <Compile Include="Converters\IntConverterTests.cs" />
     <Compile Include="Converters\FloatConverterTests.cs" />
     <Compile Include="Converters\DoubleConverterTests.cs" />
-    <Compile Include="JsonCustomTypeTests.cs" />
+    <Compile Include="Configuration\JsonCustomTypeTests.cs" />
     <Compile Include="JsonUnitTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
