+++ nanoFramework.Json/nanoFramework.Json.nfproj
@@ -34,7 +34,7 @@
     <Compile Include="Converters\BoolConverter.cs" />
     <Compile Include="Converters\ByteConverter.cs" />
     <Compile Include="Converters\CharConverter.cs" />
-    <Compile Include="Converters\ConvertersMapping.cs" />
+    <Compile Include="Configuration\ConvertersMapping.cs" />
     <Compile Include="Converters\DateTimeConverter.cs" />
     <Compile Include="Converters\DictionaryEntryConverter.cs" />
     <Compile Include="Converters\DoubleConverter.cs" />
@@ -85,9 +85,6 @@
       <HintPath>..\packages\nanoFramework.System.IO.Streams.1.1.15\lib\System.IO.Streams.dll</HintPath>
       <Private>True</Private>
     </Reference>
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Configuration\" />
   </ItemGroup>
   <Import Project="$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets" Condition="Exists('$(NanoFrameworkProjectSystemPath)NFProjectSystem.CSharp.targets')" />
   <ProjectExtensions>
